authorization and search functionality

+++ backend/task_management_system_app/views.py
@@ -1,13 +1,18 @@
 from django.shortcuts import get_object_or_404
-from rest_framework import status # type: ignore
+from rest_framework import status,generics # type: ignore
 from rest_framework.response import Response
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework_simplejwt.tokens import RefreshToken
 from .models import Category, Task
-from .serializers import CategorySerializer, TaskSerializer, RegisterSerializer, LoginSerializer
+from .serializers import CategorySerializer, TaskSerializer, RegisterSerializer, LoginSerializer, UserSerializer
 from rest_framework.decorators import action
 from django.db.models import Q
+from .permissions import IsAdminUserOrReadOnly 
+from django.contrib.auth import get_user_model
+
+User=get_user_model()
+
 
 # Authentication Views
 class RegisterAPIView(APIView):
@@ -41,12 +46,21 @@ class LogoutAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request):
-        if hasattr(request.auth,'blacklist'):
-            request.auth.blacklist()
-        # if request.auth:
-        #     token = request.auth
-        #     token.blacklist()
-        return Response({'message': 'Logout successfully'}, status=status.HTTP_204_NO_CONTENT)
+        try:
+            refresh_token = request.data.get('refresh')
+            token =  RefreshToken(refresh_token)
+            token.blacklist()
+            return Response({'message': 'Logout successfully'}, status=status.HTTP_204_NO_CONTENT)
+        except Exception as e:
+            return Response({'error': 'Invalid Token'}, status=status.HTTP_400_BAD_REQUEST)
+
+
+class UserListAPIView(generics.ListAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+    permission_classes = [IsAuthenticated]
+    
+    #needs to be completed
 
 
 # Category Views
@@ -59,6 +73,8 @@ class CategoryListCreateAPIView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
+        # if not request.user_staff:
+        #     return Response({'message':'only admin has access to create category'}, status=status.HTTP_403_FORBIDDEN)
         serializer = CategorySerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -68,7 +84,7 @@ class CategoryListCreateAPIView(APIView):
 
 
 class CategoryDetailAPIView(APIView):
-    permission_classes = [AllowAny]
+    permission_classes = [IsAuthenticated, IsAdminUserOrReadOnly]
 
     def get(self, request, pk):
         category = get_object_or_404(Category, pk=pk)
@@ -76,6 +92,8 @@ class CategoryDetailAPIView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def put(self, request, pk):
+        if not request.user_staff:
+            return Response({'message':'only admin has access to update category'}, status=status.HTTP_403_FORBIDDEN)
         category = get_object_or_404(Category, pk=pk)
         serializer = CategorySerializer(instance=category, data=request.data)
         if serializer.is_valid():
@@ -84,6 +102,9 @@ class CategoryDetailAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        # if not request.user_staff:
+        #     return Response({'message':'Only admin can delete the the category'}, status=status.HTTP_403_FORBIDDEN)
+         
         category = get_object_or_404(Category, pk=pk)
         category.delete()
         return Response({'message': 'Category deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
@@ -91,24 +112,30 @@ class CategoryDetailAPIView(APIView):
 
 # Task Views
 class TaskListCreateAPIView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsAdminUserOrReadOnly]
 
     def get(self, request):
-        tasks = Task.objects.all()
+        query = request.query_params.get('searchTerm','')
+        if query:
+            tasks = Task.objects.filter( Q(title__icontains=query) | Q(category__name__icontains=query))
+        else:
+            tasks = Task.objects.all()
         serializer = TaskSerializer(tasks, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
+        if not request.user.is_staff:
+            return Response({'message': "Only admin user is allowed to create task"}, status=status.HTTP_403_FORBIDDEN)
         serializer = TaskSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(user=request.user)
             return Response({'message': 'Task created successfully', 'task': serializer.data}, 
                             status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class TaskDetailAPIView(APIView):
-    permission_classes = [AllowAny]
+    permission_classes = [IsAuthenticated, IsAdminUserOrReadOnly]
 
     def get(self, request, pk):
         task = get_object_or_404(Task, pk=pk)
@@ -116,24 +143,28 @@ class TaskDetailAPIView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def put(self, request, pk):
-        task = get_object_or_404(Task, pk=pk)
-        serializer = TaskSerializer(instance=task, data=request.data)
+        task =get_object_or_404(Task, pk=pk)
+        if not request.user.is_staff:
+            if 'completed' in request.data:
+                task.completed=request.data['completed']
+                task.save()
+                return Response({'message':'task marked as completed'}, status=status.HTTP_200_OK)
+            return Response({'error':'non-admin user can restriction to check mark only'}, status=status.HTTP_403_FORBIDDEN)
+
+        serializer = TaskSerializer(instance=task, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        if not request.user.is_staff:
+            return Response({'message':'Only admin can delete the task'}, status=status.HTTP_403_FORBIDDEN)
+
         task = get_object_or_404(Task, pk=pk)
         task.delete()
         return Response({'message': 'Task deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
 
-    # @action(detail=False, methods=['get'], url_path='search')
-    def search(self, request):
-        query = request.query_params.get('q','')
-        results = Task.objects.filter(
-            Q(title_icontains=query) | Q(category_icontains=query)
-        )
-        serializer= TaskSerializer(results, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
     
+    
+
